Encrypt some personal data (for the first time)
Problem: according to PII guidelines, we need personal data of the users
to be stored encrypted.

Solution: bring client methods for working with passetto server.
Add utilities for storing encrypted data in Beam tables.

We distinquish two ways of storing encrypted data in table:
1. Just encrypted field.
2. Encrypted + hashed field, in order to support field lookup. This is
yet to be revised, the current implementation is not secure.

This commit also encrypts some PII in existing tables.