Access tables within a per-application schema
Our preference is for our tables to be namespaced by a per-application
schema, so let's make that happen.

The setter on the `dbEntitySchema` is not present in the version of Beam
we are using, so we should update that at some point and drop the
duplicated code.

It would also be nice to not manually add it to each field, and instead
apply it via a Traversal.