test: Spec: Add end to end API flow integration test
We currently only test the success path where the transporter back end
confirms a request after search and the consumer confirms the ride on
application back end for the confirm ride API flow. Test for health
check and list case API has also been added.

The application and transporter back end are started on their default
ports. We do not set the PORT environment variable for them to pick up.

While at it, we have split the build into library and executable. This
makes it easier to use the application level defines in tests. We drop
the application level tests in the process, as they are not very useful
and are also either commented out or do not build at the moment. The
contribution for this part comes from Arun Raghavan.

From/ToJSON and From/ToHttpApiData instances are added as required for
making the client work.

BKN-126

Signed-off-by: Sanchayan Maity <sanchayan.maity@juspay.in>