app-backend: Product: Scope queries by Person corresponding to token
Currently API handlers do not check, if the request can be made by the
user making the request. For example, if we have a Cancel case request,
the case id in this request should correspond to the registration token
of the user making the request. Person Y should not be able to cancel a
case for Person X.

Make the servant auth mechanism now return a Person and use this in
newly defined findIdByPerson to check if the case id being requested
corresponds to the token.

BKN-142

Signed-off-by: Sanchayan Maity <sanchayan.maity@juspay.in>