test: Fix integration test
Fix the list case API test. We cannot pass a hard coded registration
token for the same anymore, so start using a proper registration flow
now.

While at it, also do the following

1. The runClient function was duplicated in every test, move it to
Fixtures.
2. Start server functionality was also duplicated in each test. We now
move it to fixtures and run it at the top level, since we only really
need to do this once and stop them at the end of complete test run.

Signed-off-by: Sanchayan Maity <sanchayan.maity@juspay.in>