gateway: Support multiple apps and providers
First, we add auth to both search and on_search endpoints to allow for
some form of auth for the app and provider respectively.

Next, we look up a (currently static) set of providers on search, and
broadcast to all of them. At this point, we store a map of message_id ->
calling app.

Finally, when an on_search callback occurs, we look up the app
corresponding to the message_id and proxy the callback there.