Bring neat interface back with `(:>|)`
Problem: on the one hand, it is common for us to write many at endpoints
separated with `:<|>`, this is pretty convenient.
On the other hand, sometimes they have a common prefix and it is
tempting to e.g. write `AuthHeader :> (... :<|> ... :<|> ...)`, but it
is not convenient to work with this thing in server and client handlers;
I had this experience on the previous project and it wasn't pleasant.

Solution: for the sake of prettiness, introduce `:>|` which is a
shortcut for the latter thing. I wanted to give it a try for a long
time, and there seems to be no obvious drawbacks in it (expect for that
it may turn useless when we switch to servant-generics).