Pass auth implementation via context
Problem: previously I have added `AuthToken'` servant combinator which
allows applying some verification logic to an endpoint. But it was
designed wrong: API should not concern with implementation details,
while the current approach does not allow declaring an authentication
method without providing an implementation to it.

Soon we want to move API to common modules, and this design flaw is a
severe obstacle to that purpose.

Solution: pass the implementation via servant context instead.

Now `VerificationMethod` typeclass declares only documentation and type
of the thing which will be passed to the endpoint.
The latter also partly concerns implementation details, but I think that
if we allow the same authentication method to pass different types of
data to endpoints (so that e.g. `VerifyAdminToken` could provide either
`OrganizationId` or `PersonId`), then probably factoring authentication
out is not worth that, without servant tricks it would result in very
lot of fuss with helpers.