Refactor auth'ed search API to be common across all modules
This makes the Search API types be auth'ed and shared across all
backends, which is desirable as we want all types to be working off a
single definition of the API specification.

The actual verification method changes from backend to backend (as the
API key's meaning is different in each context).