app-backend: Clean up conversion functions for beckn spec to our own types & vice versa
The utility functions defined in Utils.Common for coverting from beckn
spec to our own types and back were ugly with from/toXToY style naming.
Use a type class for writing instances of all types we need to convert
between for cleaning this up. BecknSpecIso is representative of the
isomorphism between the two.

BKN-266

Signed-off-by: Sanchayan Maity <sanchayan.maity@juspay.in>