Poll server for async action result in tests
Problem: we have test scenarios where we e.g. run search request, and
then immediatelly ask it for results, and this does not always work as
our requests are asynchronous in their nature and so can take arbitrary
amount of time after the request has finished.

Solution: one option would be adding hardcoded waits, but this seems to
be not the best way. For instance, if I set a 0.1 sec delay, that makes
tests take significatly longer (test suites have to be executed
sequentially), while it still does not guarantee that CI or some local
PC under load would manage to execute the server's background work in
time.

The proper solution seems to be staying close to the real-life
scenario and poll the server for results until getting some non-empty
result.
This approach is implemented in this commit.