gateway: Allow graceful shutdown of gateway
We may have a search coming in, which is asynchronously forwarded to a
number of players downstream, if we get shutdown in the middle of this,
we shouldn't drop the half-done things -- so allow the existing
connections time to get handled and then shutdown.

We use a transactional variable to track the active number of
connections and when we should shutdown by detecting SIGTERM.

We also add a /healthz check for liveness that is served during
shutdown, so Kubernetes doesn't kill us off during this period.

BKN-348

Signed-off-by: Sanchayan Maity <sanchayan.maity@juspay.in>