test: Added infrastructure for FMD integration tests
The tests spin up a new test BAP server for verification of callbacks from gateway and FMD wrapper. The approach to spin up the server follows servant documentation https://docs.servant.dev/en/stable/cookbook/testing/Testing.html#how-to-test-servant-applications, adapting some code from warp (withApplication) to specify the port the server should run on (in principle, this part could also be submitted upstream). This way, we know when the server is started, and it is also shutdown gracefully when the test finishes. We use withApplication and not testWithApplication as we don't need re-throwing of server exceptions to the test code.

The test BAP uses port 8019 to avoid conflicts with other servers, and this needs creation of a new organizations in gateway and FMD wrapper.

MVars are used to wait for callbacks, and TVars are used for aggregating multiple search results (currently they come from FMD Wrapper and mock-provider-backend).
Token headers are added manually to the API so we can check what was passed in the tests.

Some refactoring is also done to reuse similar code between the tests.