backend/refactor: minor fixes required after rebase
Adds the small changes and minor fixes needed across packages
to make the project compile after doing a complete rebase iteration.

These changes occur due to new files added between the previous
rebase and the new one or are introduced as a result of
resolving merge conflicts when rebasing.

The changes include:-

* Fixing imports hiding clauses
* Adding some language pragmas
* Adding some or qualifying some imports etc.