backend/workaround: leave out buggy "scheduler-example-scheduler-exe" from run-mobility-stack-dev
There seems to be a issue in "scheduler-example-scheduler-exe" unrelated
to the GHC 9.2 upgrade.

It causes the process to consume extremely high amounts of CPU, choking it.
It also write GBs to logs for a run of few seconds/minutes.

This is being looked at and will probably be fixed or removed.
We disable it for now as a workaround untill a fix lands in main branch.