Backend/GHC-Upgrade: Mobility GHC927 Upgrade Final (#4380)
## Type of Change
<!-- Put an `x` in the boxes that apply -->

- [ ] Bugfix
- [x] New feature
- [x] Enhancement
- [x] Refactoring
- [x] Dependency updates

## Description
This PR enables the nammayatri/Backend repository to build with GHC
9.2.7 and all its upgraded package-set and dependencies that come with
the upgraded GHC927 package set, updated and rebased to work with the
newest commit.

This is part of the larger efforts to upgrade and bring the entire
nammayatri ecosystem on GHC 9.2.

To give a brief summary of changes, including but not limited to :-

- Modifies the codebase to accomodate the updated juspay and internal
dependencies, which were in-turn patched to be able to build with GHC
9.2.7
- Handles and fixes all breaking changes across major version of
repositories like Aeson v2+, Text 2+ etc.
- Updates the Nix setup accordingly.
- It also upgrades the project tooling like Haskell-Language-Server
(HLS) etc to new major versions.

### Additional Changes

- [ ] This PR modifies the database schema (database migration added)
- [ ] This PR modifies dhall configs/environment variables
- [ ] This PR contains API breaking changes

modifies the `flake.nix` and `flake.lock` for dependency upgrades.

## Motivation and Context
GHC 9.2 upgrade. A full document of motivations and benefits is being
prepared and will be shared internally across teams. It was about time.

## How did you test it?
- [x] The project builds successfully on my local
- [x] Passes all pre-commit checks

Local rideflow testing was also done, which it successfully passed.

## Checklist
<!-- Put an `x` in the boxes that apply -->

- [ ] I formatted the code and addressed linter errors
`./dev/format-all-files.sh`
- [ ] I reviewed submitted code
- [ ] I added unit tests for my changes where possible
- [ ] I added a [CHANGELOG](/CHANGELOG.md) entry if applicable
- [ ] No leak detected in leakcanary