ci: Isolate macOS Nix builds so they are faster (#9667)
We now run github-runners on the Mac Studio as well.

This seeks to resolve #9662 by doing direct local builds on Mac studio
first, followed by doing regular remote builds on NixOS (which would
download binaries from the mac to the local Nix store) so as to keep the
cache up to date.

Also, this change enables macOS builds for all PRs.